to combine two version api togetjer in a document.
